refactor game
