+++ assets/blocks/light_blue.png
