+++ game/colors.py
@@ -1,0 +1,15 @@
+from enum import Enum
+
+
+GRID_LINE_COLOR = (30, 37, 72)
+BACKGROUND_COLOR = (39, 42, 83)
+SCORE_COLOR = (255, 255, 255)
+
+class Color(Enum):
+    BLUE = 0
+    GREEN = 1
+    LIGHT_BLUE = 2
+    ORANGE = 3
+    PURPLE = 4
+    RED = 5
+    YELLOW = 6
