+++ game/fonts.py
@@ -1,0 +1,6 @@
+import pygame
+
+pygame.init()
+
+FONT = pygame.font.Font("./assets/lato_black.ttf", 90)
+UI_FONT = pygame.font.Font("./assets/lato_black.ttf", 30)
